requirements.txt
absl-py==2.3.0
astunparse==1.6.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
contourpy==1.3.3
cupy-cuda12x==13.6.0
cycler==0.12.1
dm-tree==0.1.9
fastrlock==0.8.3
filelock==3.13.1
fonttools==4.60.1
fsspec==2024.6.1
gast==0.6.0
hf-xet==1.1.3
huggingface-hub==0.32.4
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.9
llvmlite==0.44.0
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib==3.10.7
mpmath==1.3.0
networkx==3.3
noise==1.2.2
numba==0.61.2
numpy==2.1.2
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-dali-cuda120==1.50.0
nvidia-nccl-cu12==2.21.5
nvidia-nvcomp-cu12==4.2.0.14
nvidia-nvimgcodec-cu12==0.5.0.13
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvjpeg-cu12==12.4.0.16
nvidia-nvjpeg2k-cu12==0.8.1.40
nvidia-nvtiff-cu12==0.5.0.67
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.8
pycuda==2025.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytools==2025.1.6
pytz==2025.2
PyYAML==6.0.2
requests==2.32.3
safetensors==0.5.3
scipy==1.16.1
siphash24==1.7
six==1.17.0
sympy==1.13.1
timm==1.0.19
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.4.0
wrapt==1.17.2
